Enable servers to run on Windows OS.
