+++ windows/installer/cody/backend.wxs
@@ -6,8 +6,15 @@
         </Feature>
 
         <ComponentGroup Id="Backend" Directory="INSTALLDIR">
-            <Component>
+            <Component Id="Backend" Guid="{5B5A389E-C48E-45CE-A2BB-0772F0B72D0A}">
                 <File Source="bin/$(var.Platform)/signed/sourcegraph-backend.exe" />
+            </Component>
+
+            <!-- GCC components required by CGO -->
+            <Component Id="Runtime" Guid="{7AF7C689-F490-4EE2-B0B9-1ADEE1B78CEE}">
+                <File Source="c:\msys64\ucrt64\bin\libgcc_s_seh-1.dll" />
+                <File Source="c:\msys64\ucrt64\bin\libstdc++-6.dll" />
+                <File Source="c:\msys64\ucrt64\bin\libwinpthread-1.dll" />
             </Component>
         </ComponentGroup>
     </Fragment>
